refactor(PageElement): split out standalone methods to trivial base
